+++ src/EventGrid/EventGrid/Domain/GetAzureEventGridDomain.cs
@@ -12,6 +12,7 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using System;
 using System.Collections.Generic;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.EventGrid.Models;
@@ -31,7 +32,8 @@ namespace Microsoft.Azure.Commands.EventGrid
         VerbsCommon.Get,
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "EventGridDomain",
         DefaultParameterSetName = ResourceGroupNameParameterSet),
-    OutputType(typeof(PSDomain), typeof(PSDomainListInstance))]
+    OutputType(typeof(PSDomain),
+    typeof(PSDomainListInstance))]
 
     public class GetAzureEventGridDomain : AzureEventGridCmdletBase
     {
@@ -70,10 +72,57 @@ namespace Microsoft.Azure.Commands.EventGrid
         [ValidateNotNullOrEmpty]
         public string ResourceId { get; set; }
 
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = DomainNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = ResourceGroupNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = ResourceIdEventSubscriptionParameterSet)]
+        [ValidateNotNullOrEmpty]
+        public string ODataQuery { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.TopHelp,
+            ParameterSetName = DomainNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.TopHelp,
+            ParameterSetName = ResourceGroupNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.TopHelp,
+            ParameterSetName = ResourceIdEventSubscriptionParameterSet)]
+        [ValidateRange(1, 100)]
+        public int? Top { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.NextLinkHelp,
+            ParameterSetName = NextLinkParameterSet)]
+        [ValidateNotNullOrEmpty]
+        public string NextLink { get; set; }
+
         public override void ExecuteCmdlet()
         {
             string resourceGroupName = string.Empty;
             string domainName = string.Empty;
+            IEnumerable<Domain> domainsList;
+            string nextLink = null;
+            string newNextLink = null;
 
             if (!string.IsNullOrEmpty(this.ResourceId))
             {
@@ -89,8 +138,31 @@ namespace Microsoft.Azure.Commands.EventGrid
             {
                 resourceGroupName = this.ResourceGroupName;
             }
+            else if (!string.IsNullOrEmpty(this.NextLink))
+            {
+                // Other parameters should be null or ignored if nextLink is specified.
+                nextLink = this.NextLink;
+            }
 
-            if (!string.IsNullOrEmpty(resourceGroupName) && !string.IsNullOrEmpty(domainName))
+            if (!string.IsNullOrEmpty(nextLink))
+            {
+                // Get Next page of domains. Get the proper next API to be called based on the nextLink.
+                Uri uri = new Uri(nextLink);
+                string path = uri.AbsolutePath;
+
+                if (path.IndexOf("/resourceGroups/", StringComparison.OrdinalIgnoreCase) != -1)
+                {
+                    (domainsList, newNextLink) = this.Client.ListDomainsByResourceGroupNext(nextLink);
+                }
+                else
+                {
+                    (domainsList, newNextLink) = this.Client.ListDomainBySubscriptionNext(nextLink);
+                }
+
+                PSDomainListPagedInstance pSDomainListPagedInstance = new PSDomainListPagedInstance(domainsList, newNextLink);
+                this.WriteObject(pSDomainListPagedInstance, true);
+            }
+            else if (!string.IsNullOrEmpty(resourceGroupName) && !string.IsNullOrEmpty(domainName))
             {
                 // Get details of the Event Grid domain
                 Domain domain = this.Client.GetDomain(resourceGroupName, domainName);
@@ -100,28 +172,16 @@ namespace Microsoft.Azure.Commands.EventGrid
             else if (!string.IsNullOrEmpty(resourceGroupName) && string.IsNullOrEmpty(domainName))
             {
                 // List all Event Grid domains in the given resource group
-                IEnumerable<Domain> domainsList = this.Client.ListDomainsByResourceGroup(resourceGroupName);
-
-                List<PSDomainListInstance> psDomainsList = new List<PSDomainListInstance>();
-                foreach (Domain domain in domainsList)
-                {
-                    psDomainsList.Add(new PSDomainListInstance(domain));
-                }
-
-                this.WriteObject(psDomainsList, true);
+                (domainsList, newNextLink) = this.Client.ListDomainsByResourceGroup(resourceGroupName, this.ODataQuery, this.Top);
+                PSDomainListPagedInstance pSDomainListPagedInstance = new PSDomainListPagedInstance(domainsList, newNextLink);
+                this.WriteObject(pSDomainListPagedInstance, true);
             }
-            else
+            else if (string.IsNullOrEmpty(resourceGroupName) && string.IsNullOrEmpty(domainName))
             {
                 // List all Event Grid domains in the given subscription
-                IEnumerable<Domain> domainsList = this.Client.ListDomainsBySubscription();
-
-                List<PSDomainListInstance> psDomainsList = new List<PSDomainListInstance>();
-                foreach (Domain domain in domainsList)
-                {
-                    psDomainsList.Add(new PSDomainListInstance(domain));
-                }
-
-                this.WriteObject(psDomainsList, true);
+                (domainsList, newNextLink) = this.Client.ListDomainsBySubscription(this.ODataQuery, this.Top);
+                PSDomainListPagedInstance pSDomainListPagedInstance = new PSDomainListPagedInstance(domainsList, newNextLink);
+                this.WriteObject(pSDomainListPagedInstance, true);
             }
         }
     }