+++ src/EventGrid/EventGrid/Domain/GetAzureEventGridDomainTopic.cs
@@ -12,6 +12,7 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using System;
 using System.Collections.Generic;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.EventGrid.Models;
@@ -30,7 +31,8 @@ namespace Microsoft.Azure.Commands.EventGrid
         VerbsCommon.Get,
         ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "EventGridDomainTopic",
         DefaultParameterSetName = DomainTopicNameParameterSet),
-    OutputType(typeof(PSDomainTopic), typeof(PSDomainTopicListInstance))]
+    OutputType(typeof(PSDomainTopic),
+    typeof(PSDomainTopicListInstance))]
 
     public class GetAzureEventGridDomainTopic : AzureEventGridCmdletBase
     {
@@ -52,6 +54,7 @@ namespace Microsoft.Azure.Commands.EventGrid
             HelpMessage = EventGridConstants.DomainNameHelp,
             ParameterSetName = DomainTopicNameParameterSet)]
         [ValidateNotNullOrEmpty]
+        [Alias(AliasDomain)]
         public string DomainName { get; set; }
 
         [Parameter(
@@ -60,51 +63,82 @@ namespace Microsoft.Azure.Commands.EventGrid
             HelpMessage = EventGridConstants.DomainTopicNameHelp,
             ParameterSetName = DomainTopicNameParameterSet)]
         [ValidateNotNullOrEmpty]
-        public string DomainTopicName { get; set; }
+        [Alias(AliasDomainTopicName)]
+        public string Name { get; set; }
 
         [Parameter(
             Mandatory = true,
             ValueFromPipelineByPropertyName = true,
             Position = 0,
             HelpMessage = EventGridConstants.DomainOrDomainTopicResourceIdHelp,
-            ParameterSetName = ResourceIdEventSubscriptionParameterSet)]
+            ParameterSetName = ResourceIdDomainTopicParameterSet)]
         [ValidateNotNullOrEmpty]
         public string ResourceId { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = DomainTopicNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = ResourceIdDomainTopicParameterSet)]
+        [ValidateNotNullOrEmpty]
+        public string ODataQuery { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = DomainTopicNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = ResourceIdDomainTopicParameterSet)]
+        [ValidateRange(1, 100)]
+        public int? Top { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.NextLinkHelp,
+            ParameterSetName = NextLinkParameterSet)]
+        [ValidateNotNullOrEmpty]
+        public string NextLink { get; set; }
 
         public override void ExecuteCmdlet()
         {
             string resourceGroupName = string.Empty;
             string domainName = string.Empty;
             string domainTopicName = string.Empty;
+            string nextLink = this.NextLink;
+            string newNextLink = null;
+            IEnumerable<DomainTopic> domainTopicsList;
+            List <PSDomainTopicListInstance> psDomainTopicsList = new List<PSDomainTopicListInstance>();
 
             if (!string.IsNullOrEmpty(this.ResourceId))
             {
                 EventGridUtils.GetResourceGroupNameAndDomainNameAndDomainTopicName(this.ResourceId, out resourceGroupName, out domainName, out domainTopicName);
             }
-            else if (!string.IsNullOrEmpty(this.DomainName))
+            else
             {
-                // If Domain name is provided, ResourceGroup should be non-empty as well
                 resourceGroupName = this.ResourceGroupName;
                 domainName = this.DomainName;
-
-                if (!string.IsNullOrEmpty(this.DomainTopicName))
-                {
-                    domainTopicName = this.DomainTopicName;
-                }
-            }
-            else if (!string.IsNullOrEmpty(this.DomainTopicName))
-            {
-                // If Domain topic name is provided, ResourceGroup and domain name should be non-empty as well
-                resourceGroupName = this.ResourceGroupName;
-                domainName = this.DomainName;
-                domainTopicName = this.DomainTopicName;
-            }
-            else if (!string.IsNullOrEmpty(this.ResourceGroupName))
-            {
-                resourceGroupName = this.ResourceGroupName;
+                domainTopicName = this.Name;
             }
 
-            if (!string.IsNullOrEmpty(resourceGroupName) && !string.IsNullOrEmpty(domainName))
+            if (!string.IsNullOrEmpty(nextLink))
+            {
+                // Get Next page of domain topics.
+                (domainTopicsList, newNextLink) = this.Client.ListDomainTopicsByDomainNext(nextLink);
+
+                PSDomainTopicListPagedInstance pSDomainTopicListPagedInstance = new PSDomainTopicListPagedInstance(domainTopicsList, newNextLink);
+                this.WriteObject(pSDomainTopicListPagedInstance, true);
+            }
+            else if (!string.IsNullOrEmpty(resourceGroupName) && !string.IsNullOrEmpty(domainName))
             {
                 if (!string.IsNullOrEmpty(domainTopicName))
                 {
@@ -116,15 +150,9 @@ namespace Microsoft.Azure.Commands.EventGrid
                 else
                 {
                     // List all Event Grid domain topics in the given resource group/domain
-                    IEnumerable<DomainTopic> domainTopicsList = this.Client.ListDomainTopicsByDomain(resourceGroupName, domainName);
-
-                    List<PSDomainTopicListInstance> psDomainTopicsList = new List<PSDomainTopicListInstance>();
-                    foreach (DomainTopic domainTopic in domainTopicsList)
-                    {
-                        psDomainTopicsList.Add(new PSDomainTopicListInstance(domainTopic));
-                    }
-
-                    this.WriteObject(psDomainTopicsList, true);
+                    (domainTopicsList, newNextLink) = this.Client.ListDomainTopicsByDomain(resourceGroupName, domainName, this.ODataQuery, this.Top);
+                    PSDomainTopicListPagedInstance pSDomainTopicListPagedInstance = new PSDomainTopicListPagedInstance(domainTopicsList, newNextLink);
+                    this.WriteObject(pSDomainTopicListPagedInstance, true);
                 }
             }
         }