+++ src/EventGrid/EventGrid/Topic/GetAzureEventGridTopic.cs
@@ -12,6 +12,7 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using System;
 using System.Collections.Generic;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.EventGrid.Models;
@@ -26,7 +27,11 @@ namespace Microsoft.Azure.Commands.EventGrid
     /// <para> If Topic name provided, a single Topic details will be returned</para>
     /// <para> If Topic name not provided, list of Topics will be returned</para>
     /// </summary>
-    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "EventGridTopic",DefaultParameterSetName = ResourceGroupNameParameterSet),OutputType(typeof(PSTopic), typeof(PSTopicListInstance))]
+    [Cmdlet("Get",
+            ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "EventGridTopic",
+            DefaultParameterSetName = ResourceGroupNameParameterSet),
+        OutputType(typeof(PSTopic),
+        typeof(PSTopicListInstance))]
     public class GetAzureRmEventGridTopic : AzureEventGridCmdletBase
     {
         [Parameter(
@@ -64,10 +69,57 @@ namespace Microsoft.Azure.Commands.EventGrid
         [ValidateNotNullOrEmpty]
         public string ResourceId { get; set; }
 
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = TopicNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = ResourceGroupNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.ODataQueryHelp,
+            ParameterSetName = ResourceIdEventSubscriptionParameterSet)]
+        [ValidateNotNullOrEmpty]
+        public string ODataQuery { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.TopHelp,
+            ParameterSetName = TopicNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.TopHelp,
+            ParameterSetName = ResourceGroupNameParameterSet)]
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.TopHelp,
+            ParameterSetName = ResourceIdEventSubscriptionParameterSet)]
+        [ValidateRange(1, 100)]
+        public int? Top { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = EventGridConstants.NextLinkHelp,
+            ParameterSetName = NextLinkParameterSet)]
+        [ValidateNotNullOrEmpty]
+        public string NextLink { get; set; }
+
         public override void ExecuteCmdlet()
         {
             string resourceGroupName = string.Empty;
             string topicName = string.Empty;
+            IEnumerable<Topic> topicsList;
+            string nextLink = null;
+            string newNextLink = null;
 
             if (!string.IsNullOrEmpty(this.ResourceId))
             {
@@ -83,8 +135,31 @@ namespace Microsoft.Azure.Commands.EventGrid
             {
                 resourceGroupName = this.ResourceGroupName;
             }
+            else if (!string.IsNullOrEmpty(this.NextLink))
+            {
+                // Other parameters should be null or ignored if nextLink is specified.
+                nextLink = this.NextLink;
+            }
 
-            if (!string.IsNullOrEmpty(resourceGroupName) && !string.IsNullOrEmpty(topicName))
+            if (!string.IsNullOrEmpty(nextLink))
+            {
+                // Get Next page of topics. Get the proper next API to be called based on the nextLink.
+                Uri uri = new Uri(nextLink);
+                string path = uri.AbsolutePath;
+
+                if (path.IndexOf("/resourceGroups/", StringComparison.OrdinalIgnoreCase) != -1)
+                {
+                    (topicsList, newNextLink) = this.Client.ListTopicsByResourceGroupNext(nextLink);
+                }
+                else
+                {
+                    (topicsList, newNextLink) = this.Client.ListTopicBySubscriptionNext(nextLink);
+                }
+
+                PSTopicListPagedInstance pSTopicListPagedInstance = new PSTopicListPagedInstance(topicsList, newNextLink);
+                this.WriteObject(pSTopicListPagedInstance, true);
+            }
+            else if (!string.IsNullOrEmpty(resourceGroupName) && !string.IsNullOrEmpty(topicName))
             {
                 // Get details of the Event Grid topic
                 Topic topic = this.Client.GetTopic(resourceGroupName, topicName);
@@ -94,28 +169,16 @@ namespace Microsoft.Azure.Commands.EventGrid
             else if (!string.IsNullOrEmpty(resourceGroupName) && string.IsNullOrEmpty(topicName))
             {
                 // List all Event Grid topics in the given resource group
-                IEnumerable<Topic> topicsList = this.Client.ListTopicsByResourceGroup(resourceGroupName);
-
-                List<PSTopicListInstance> psTopicsList = new List<PSTopicListInstance>();
-                foreach (Topic topic in topicsList)
-                {
-                    psTopicsList.Add(new PSTopicListInstance(topic));
-                }
-
-                this.WriteObject(psTopicsList, true);
+                (topicsList, newNextLink) = this.Client.ListTopicsByResourceGroup(resourceGroupName, this.ODataQuery, this.Top);
+                PSTopicListPagedInstance pSTopicListPagedInstance = new PSTopicListPagedInstance(topicsList, newNextLink);
+                this.WriteObject(pSTopicListPagedInstance, true);
             }
-            else
+            else if (string.IsNullOrEmpty(resourceGroupName) && string.IsNullOrEmpty(topicName))
             {
                 // List all Event Grid topics in the given subscription
-                IEnumerable<Topic> topicsList = this.Client.ListTopicsBySubscription();
-
-                List<PSTopicListInstance> psTopicsList = new List<PSTopicListInstance>();
-                foreach (Topic topic in topicsList)
-                {
-                    psTopicsList.Add(new PSTopicListInstance(topic));
-                }
-
-                this.WriteObject(psTopicsList, true);
+                (topicsList, newNextLink) = this.Client.ListTopicsBySubscription(this.ODataQuery, this.Top);
+                PSTopicListPagedInstance pSTopicListPagedInstance = new PSTopicListPagedInstance(topicsList, newNextLink);
+                this.WriteObject(pSTopicListPagedInstance, true);
             }
         }
     }