+++ src/EventGrid/EventGrid/ChangeLog.md
@@ -19,6 +19,21 @@
 -->
 ## Upcoming Release
 
+## Version 1.1.1
+* Updated to use the 2019-02-01-preview API version.
+* New commands:
+    - New-AzureRmEventGridDomainTopic:
+        - Creates a new Azure Event Grid Domain Topic.
+    - Remove-AzureRmEventGridDomainTopic:
+        - Remove an existing Azure Event Grid Domain Topic.
+
+* Update the following cmdlets to support new scenarios in 2019-02-01-preview API version
+    - New-AzureRmEventGridSubscription/Update-AzureRmEventGridSubscription:
+        - Add new enum for servicebusqueue as destination.
+        - Disllow usage of "All" in IncludesEventTypes option and replace it with $null
+    - Get-AzEventGridTopic, Get-AzEventGridDomain, Get-AzEventGridDomainTopic, Get-AzEventGridSubscription:
+        - Add new optional parameters (Top, ODataQuery and NextLink) to support pagination.
+
 ## Version 1.1.0
 * Updated to use the 2019-01-01 API version.
 * Update the following cmdlets to support new scenario in 2019-01-01 API version